backend/requirements.txt
aiomysql==0.2.0
aiosqlite==0.19.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.10.4
async-timeout==5.0.1
asyncmy==0.2.10
asyncpg==0.29.0
click==8.1.8
dnspython==2.7.0
exceptiongroup==1.3.1
fastapi==0.128.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
motor==3.7.1
cryptography>=41.0.0
psycopg2-binary>=2.9.9
prometheus_client>=0.19.0
psutil>=5.9.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator>=2.0.0
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
pymongo==4.15.5
PyMySQL==1.1.2
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uvicorn==0.39.0
